// frontend/src/types/Movie.ts
export interface Movie {
  show_id?: number;
  type: string;
  title: string;
  director: string;
  cast: string;
  country: string;
  release_year: number;
  rating: string;
  duration: string;
  description: string;
  action: number;
  adventure: number;
  anime: number;
  british_Docuseries: number;
  children: number;
  comedies: number;
  comedies_Dramas: number;
  comedies_International: number;
  comedies_Romantic: number;
  crime_TV: number;
  documentaries: number;
  documentaries_International: number;
  docuseries: number;
  dramas: number;
  dramas_International: number;
  dramas_Romantic: number;
  family: number;
  fantasy: number;
  horror: number;
  international_Thrillers: number;
  international_Romantic_Dramas_TV: number;
  kids_TV: number;
  language_TV: number;
  musicals: number;
  nature_TV: number;
  reality_TV: number;
  spirituality: number;
  tV_Action: number;
  tV_Comedies: number;
  tV_Dramas: number;
  talk_Shows: number;
  thrillers: number;
}